FFMpeg rendering, Initial Pattern Creation & tests
FFMpeg rendering will now be done by piping every frame into a
subprocess. There seems, however, to be some error with the subprocess
stdin buffer being filled and it hanging resulting in me having to flush
it, close the process and start a new one. A bunch of movies are created
and then concatenated at the end. This is still better than a data.txt
which easily goes above 100s if gigabytes in size, but not still not
nice. Bug in osproc?

Inside drawing.nim there has been some initial work on pattern creation.
Patterns ends the current frame and saves the pixels written by nanovg.
It then clears a portion of the screen, draws a pattern, creates an
Image and a Pain from those pixels, resets the screen to its original
state and returns the Paint/Pattern. Simple patterns are cached.

Initial unittests have been setup. Workflows works out fo the box? Probs
not...